LPS-83992 Move XML related classes to xml subpackage
