+++ modules/apps/bean-portlet-cdi-extension/src/main/java/com/liferay/bean/portlet/cdi/extension/internal/xml/LiferayDescriptor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.bean.portlet.cdi.extension.internal;
+package com.liferay.bean.portlet.cdi.extension.internal.xml;
 
 import java.util.HashMap;
 import java.util.Map;
